+++ Cargo.toml
@@ -32,6 +32,7 @@ rayon = "1.8"
 filesize = "0.2.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+bitflags = "2.6.0"
 
 [profiles.release]
 opt-level = 3
