+++ src/generate/mod.rs
@@ -1,14 +1,14 @@
 pub mod config;
 pub mod context;
 pub mod context_collection;
-pub mod cpp_type;
-pub mod cpp_type_tag;
+pub mod cs_type;
+pub mod cs_type_tag;
 pub mod cs_context_collection;
 pub mod cs_fields;
-pub mod cs_type;
 pub mod members;
-pub mod members_serialize;
 pub mod metadata;
 pub mod offsets;
 pub mod type_extensions;
 pub mod writer;
+
+pub mod cpp;
